+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Master
+
+###### Enhancements
+
+* Add a `ignore_in_command_lookup` option to commands, which makes it possible
+  to have anonymous command classes that are or only meant to provide common
+  functionality, but are otherwise completely ignored during parsing, command
+  lookup, and help banner printing.
+  [Eloy Durán](https://github.com/alloy)
+
 
 ## 0.4.0